maybe fixed the swerve steering problem?
the core thing was using the AnalogRaw measurement rather than the
Position measurement, which i think was winding up, as if the encoder
were measuring linear distance, like the drive encoder does.
i also added some feedforward to steering, anticipating that we'll
want to get rid of the overshoot, and i guessed at some reasonable
P and D values.